requirements.txt
qdrant-client==1.15.1
sentence-transformers==5.1.0
langchain==0.3.27
langchain-community==0.3.27
tavily-python==0.7.10
wikipedia==1.4.0
langchain-openai==0.3.28
streamlit==1.48.0
python-docx==1.2.0
reportlab==4.4.3
cohere==5.16.2